Added org/:id to public api rouer. Code Climate fixes
