+++ dpc-go/dpc-api/src/service/admin/router_test.go
@@ -92,7 +92,7 @@ func TestRouterTestSuite(t *testing.T) {
 
 func (suite *RouterTestSuite) TestErrorHandling() {
 
-	suite.mockOrg.On("Read", mock.Anything, mock.Anything).Once().Run(func(arg mock.Arguments) {
+	suite.mockOrg.On("Delete", mock.Anything, mock.Anything).Once().Run(func(arg mock.Arguments) {
 		w := arg.Get(0).(http.ResponseWriter)
 		r := arg.Get(1).(*http.Request)
 		fhirror.GenericServerIssue(r.Context(), w)
@@ -100,7 +100,7 @@ func (suite *RouterTestSuite) TestErrorHandling() {
 
 	ts := httptest.NewServer(suite.router)
 
-	req, _ := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/%s", ts.URL, "v2/Organization/12345"), nil)
+	req, _ := http.NewRequest(http.MethodDelete, fmt.Sprintf("%s/%s", ts.URL, "v2/Organization/12345"), nil)
 	req.Header.Set(middleware.RequestIDHeader, "54321")
 	res, _ := http.DefaultClient.Do(req)
 
@@ -124,34 +124,6 @@ func (suite *RouterTestSuite) TestErrorHandling() {
       ],
       "resourceType": "OperationOutcome"
     }`)
-}
-
-func (suite *RouterTestSuite) TestOrganizationGetRoutes() {
-	var capturedRequestID string
-	suite.mockOrg.On("Read", mock.Anything, mock.Anything).Once().Run(func(arg mock.Arguments) {
-		r := arg.Get(1).(*http.Request)
-		capturedRequestID = r.Header.Get(middleware.RequestIDHeader)
-		w := arg.Get(0).(http.ResponseWriter)
-		_, _ = w.Write(apitest.AttributionOrgResponse())
-	})
-
-	ts := httptest.NewServer(suite.router)
-
-	req, _ := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/%s", ts.URL, "v2/Organization/12345"), nil)
-	req.Header.Set(middleware.RequestIDHeader, "54321")
-	res, _ := http.DefaultClient.Do(req)
-
-	b, _ := ioutil.ReadAll(res.Body)
-	var v map[string]interface{}
-	_ = json.Unmarshal(b, &v)
-
-	assert.Equal(suite.T(), "application/fhir+json; charset=UTF-8", res.Header.Get("Content-Type"))
-	assert.Equal(suite.T(), http.StatusOK, res.StatusCode)
-	assert.Equal(suite.T(), "54321", capturedRequestID)
-	assert.NotNil(suite.T(), v)
-	assert.NotContains(suite.T(), v, "info")
-	assert.Contains(suite.T(), v, "resourceType")
-	assert.Equal(suite.T(), v["resourceType"], "Organization")
 }
 
 func (suite *RouterTestSuite) TestOrganizationPostRoute() {
