+++ dpc-go/dpc-api/src/service/public/router_test.go
@@ -1,19 +1,20 @@
 package public
 
 import (
-    "encoding/json"
-    "fmt"
-    "io/ioutil"
-    "net/http"
-    "net/http/httptest"
-    "testing"
+	"encoding/json"
+	"fmt"
+	"io/ioutil"
+	"net/http"
+	"net/http/httptest"
+	"testing"
 
-    "github.com/go-chi/chi/middleware"
-    "github.com/stretchr/testify/assert"
-    "github.com/stretchr/testify/mock"
-    "github.com/stretchr/testify/suite"
+	"github.com/CMSgov/dpc/api/apitest"
+	"github.com/go-chi/chi/middleware"
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/mock"
+	"github.com/stretchr/testify/suite"
 
-    middleware2 "github.com/CMSgov/dpc/api/middleware"
+	middleware2 "github.com/CMSgov/dpc/api/middleware"
 )
 
 type MockController struct {
@@ -66,20 +67,22 @@ func (mjc *MockSsasController) CreateSystem(w http.ResponseWriter, r *http.Reque
 
 type RouterTestSuite struct {
 	suite.Suite
-	router      http.Handler
-	mockMeta    *MockController
-	mockGroup   *MockController
-	mockData    *MockFileController
-	mockJob     *MockJobController
+	router    http.Handler
+	mockOrg   *MockController
+	mockMeta  *MockController
+	mockGroup *MockController
+	mockData  *MockFileController
+	mockJob   *MockJobController
 }
 
 func (suite *RouterTestSuite) SetupTest() {
+	suite.mockOrg = &MockController{}
 	suite.mockMeta = &MockController{}
 	suite.mockGroup = &MockController{}
 	suite.mockData = &MockFileController{}
 	suite.mockJob = &MockJobController{}
 
-	suite.router = buildPublicRoutes(suite.mockMeta, suite.mockGroup, suite.mockData, suite.mockJob)
+	suite.router = buildPublicRoutes(suite.mockOrg, suite.mockMeta, suite.mockGroup, suite.mockData, suite.mockJob)
 }
 
 func TestRouterTestSuite(t *testing.T) {
@@ -124,4 +127,32 @@ func (suite *RouterTestSuite) TestGroupExportRoute() {
 	assert.Equal(suite.T(), http.StatusAccepted, res.StatusCode)
 	assert.Equal(suite.T(), "54321", capturedRequestID)
 	assert.Nil(suite.T(), v)
+}
+
+func (suite *RouterTestSuite) TestOrganizationGetRoutes() {
+	var capturedRequestID string
+	suite.mockOrg.On("Read", mock.Anything, mock.Anything).Once().Run(func(arg mock.Arguments) {
+		r := arg.Get(1).(*http.Request)
+		capturedRequestID = r.Header.Get(middleware.RequestIDHeader)
+		w := arg.Get(0).(http.ResponseWriter)
+		_, _ = w.Write(apitest.AttributionOrgResponse())
+	})
+
+	ts := httptest.NewServer(suite.router)
+
+	req, _ := http.NewRequest(http.MethodGet, fmt.Sprintf("%s/%s", ts.URL, "v2/Organization/12345"), nil)
+	req.Header.Set(middleware.RequestIDHeader, "54321")
+	res, _ := http.DefaultClient.Do(req)
+
+	b, _ := ioutil.ReadAll(res.Body)
+	var v map[string]interface{}
+	_ = json.Unmarshal(b, &v)
+
+	assert.Equal(suite.T(), "application/fhir+json; charset=UTF-8", res.Header.Get("Content-Type"))
+	assert.Equal(suite.T(), http.StatusOK, res.StatusCode)
+	assert.Equal(suite.T(), "54321", capturedRequestID)
+	assert.NotNil(suite.T(), v)
+	assert.NotContains(suite.T(), v, "info")
+	assert.Contains(suite.T(), v, "resourceType")
+	assert.Equal(suite.T(), v["resourceType"], "Organization")
 }
